+++ includes/CMB2_Sanitize.php
@@ -190,16 +190,18 @@ class CMB2_Sanitize {
 	/**
 	 * Converts text date to timestamp
 	 * @since  1.0.2
-	 * @return string       Timestring
+	 * @return string Timestring
 	 */
 	public function text_date_timestamp() {
-		return is_array( $this->value ) ? array_map( 'strtotime', $this->value ) : strtotime( $this->value );
+		return is_array( $this->value )
+			? array_map( array( $this->field, 'get_timestamp_from_value' ), $this->value )
+			: $this->field->get_timestamp_from_value( $this->value );
 	}
 
 	/**
 	 * Datetime to timestamp
 	 * @since  1.0.1
-	 * @return string       Timestring
+	 * @return string Timestring
 	 */
 	public function text_datetime_timestamp( $repeat = false ) {
 
@@ -212,10 +214,12 @@ class CMB2_Sanitize {
 			return $repeat_value;
 		}
 
-		$this->value = strtotime( $this->value['date'] . ' ' . $this->value['time'] );
+		if ( isset( $this->value['date'], $this->value['time'] ) ) {
+			$this->value = $this->field->get_timestamp_from_value( $this->value['date'] . ' ' . $this->value['time'] );
+		}
 
 		if ( $tz_offset = $this->field->field_timezone_offset() ) {
-			$this->value += $tz_offset;
+			$this->value += (int) $tz_offset;
 		}
 
 		return $this->value;
@@ -260,9 +264,13 @@ class CMB2_Sanitize {
 			$tzstring = false !== $tzstring ? $tzstring : timezone_name_from_abbr( '', 0, 0 );
 		}
 
+		$full_format = $this->field->args['date_format'] . ' ' . $this->field->args['time_format'];
+		$full_date   = $this->value['date'] . ' ' . $this->value['time'];
+
 		try {
-			$this->value = new DateTime( $this->value['date'] . ' ' . $this->value['time'], new DateTimeZone( $tzstring ) );
-			$this->value = serialize( $this->value );
+			$datetime = date_create_from_format( $full_format, $full_date );
+			$datetime->setTimezone( new DateTimeZone( $tzstring ) );
+			$this->value = serialize( $datetime );
 		} catch ( Exception $e ) {
 			cmb2_utils()->log_if_debug( __METHOD__, __LINE__, $e->getMessage() );
 		}